lots of updates!

+++ app/routers/weights.py
@@ -1,5 +1,7 @@
+from typing import Annotated
+
 from fastapi import APIRouter, Depends, Query, HTTPException
-from sqlmodel import Session, select
+from sqlmodel import Session, select, desc
 
 from app.db.models import (
     Weight,
@@ -7,6 +9,7 @@ from app.db.models import (
     WeightCreate,
 )
 from app.dependencies import get_session
+from app.auth import get_current_user, require_permission
 
 router = APIRouter(
     prefix="/weights",
@@ -15,9 +18,18 @@ router = APIRouter(
 
 
 @router.post("/", response_model=WeightRead)
+@require_permission("create_all_weights", "create_own_weight")
 def create_weight(
-    *, session: Session = Depends(get_session), weight: WeightCreate
+    *,
+    session: Session = Depends(get_session),
+    weight: WeightCreate,
+    current_user: Annotated[str, Depends(get_current_user)],
 ):
+    if weight.user_id != current_user.id and not current_user.has("create_all_weights"):
+        raise HTTPException(
+            status_code=400,
+            detail="User does not have permission to create weights for another user",
+        )
     db_weight = Weight.from_orm(weight)
     session.add(db_weight)
     session.commit()
@@ -26,17 +38,25 @@ def create_weight(
 
 
 @router.get("/", response_model=list[WeightRead])
+@require_permission("read_all_weights", "read_own_weight")
 def read_weights(
     *,
     session: Session = Depends(get_session),
     offset: int = 0,
     limit: int = Query(default=100, lte=100),
+    user_id: int | None = None,
+    current_user: Annotated[str, Depends(get_current_user)],
 ):
+    if user_id != current_user.id and not current_user.has("read_all_weights"):
+        raise HTTPException(
+            status_code=400,
+            detail="User does not have permission to read weights for another user",
+        )
+    query = select(Weight)
+    if user_id:
+        query = query.where(Weight.user_id == user_id)
     return session.exec(
-        select(Weight)
-        .order_by(Weight.id)
-        .offset(offset)
-        .limit(limit)
+        query.order_by(desc(Weight.date)).offset(offset).limit(limit)
     ).all()
 
 
@@ -44,10 +64,21 @@ def read_weights(
     "/{weight_id}",
     response_model=WeightRead,
 )
+@require_permission("read_all_weights", "read_own_weight")
 def read_weight(
-    *, session: Session = Depends(get_session), weight_id: int
+    *,
+    session: Session = Depends(get_session),
+    weight_id: int,
+    current_user: Annotated[str, Depends(get_current_user)],
 ):
     if weight := session.get(Weight, weight_id):
+        if weight.user_id != current_user.id and not current_user.has(
+            "read_all_weights"
+        ):
+            raise HTTPException(
+                status_code=400,
+                detail="User does not have permission to read weights for another user",
+            )
         return weight
     else:
         raise HTTPException(status_code=404, detail="Weight not found")
@@ -85,12 +116,21 @@ def read_weight(
 
 
 @router.delete("/{weight_id}")
+@require_permission("delete_all_weights", "delete_own_weight")
 def delete_weight(
-    *, session: Session = Depends(get_session), weight_id: int
+    *,
+    session: Session = Depends(get_session),
+    weight_id: int,
+    current_user: Annotated[str, Depends(get_current_user)],
 ):
     weight = session.get(Weight, weight_id)
     if not weight:
         raise HTTPException(status_code=404, detail="Weight not found")
+    if weight.user_id != current_user.id and not current_user.has("delete_all_weights"):
+        raise HTTPException(
+            status_code=400,
+            detail="User does not have permission to delete weights for another user",
+        )
     session.delete(weight)
     session.commit()
     return {"ok": True}
